requirements.txt
agora-token-builder==1.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.146
botocore==1.29.146
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-phonenumber-field==7.1.0
django-shell-plus==1.1.7
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.5
phonenumberslite==8.13.11
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==1.26.16
zope.interface==6.0
